requirement.txt

Django==2.0.4
django-allauth==0.32.0
pillow==5.1.0